Remove old les code, fix food
